suite: handle variable number of args
If we turn suite into a case-lambda, we can handle the absence or
presence of options, setups, and teardowns. The minimum required
arguments are now the suite description and a list of tests. This makes
life easier for test writers at the cost of having to write tests
_before_ options, setups, and teardowns. However, I think this is a good
thing because it forces them to focus on tests _first_, and _then_
everything else.