run-file: start adding an automated test runner
This function can run all the suites in a single file. The next step is
to write a function to call run-file for all files in a certain
subdirectory.

Also, change the suite function to return results immediately instead of
stashing its arguments for later use and then returning a lambda that
returns the results.