output-normal: reimplement around new sets of kw args
Based on the new combinations of keyword arguments being passed in to
the 'output-cb' function(s), we can implement 'output-normal' as a
minimalistic-but-useful output format. The implementation is just a
matter of checking for the presence or absence of sets of keyword
arguments, alone or in combination with each other.