output-normal: fix error diagnostics output
The problem is the 'if-let' syntax. If 'if-let' is passed a test
expression that evaluates to a #f value, then 'if-let' will skip the
'then' expression and just evaluate the 'else' expression. To avoid
that, we don't put any plain booleans into anything that might get
evaluated by 'if-let'. Main case (in point), into any of the keyword
arguments to 'output-normal'.