suite: add option to output tally after each suite
Adding this as an option that can be given to a suite in a test file,
but not at the command line, because it doesn't make sense for every
suite to output a tally, only for the ggspec test runner to output a
final tally after all the suites have finished. This might change in
future--the capability is now there.