suite-add-option: failing test for duplicate, conflicting options
We shouldn't add an option to a suite if it already has that option, in
the code definition of the suite. In other words, options defined in the
code should have a higher precedence than options passed in at the
command line. This affects things like colouring output and choosing the
output format.