run-tree: fix bug for non-Scheme files in 'spec' dir
The bug was that the path (file) processing function inside the
'file-system-fold' call would return an unspecified value if it came
across a non-Scheme (*.scm) file in the 'spec' directory. The fix is to
return the existing result list unchanged when it does so.

Using the new 'results-add' function helped to expose this bug.